pyproject.toml
[tool.poetry]
name = "discord-embed-preview"
version = "0.1.0"
description = "Discord message link preview bot"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false
repository = "https://github.com/yourusername/discord-embed-preview"
keywords = ["discord", "bot", "preview", "embed"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: Japanese",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.3.2"
python-dotenv = "^1.0.0"
Pillow = "^9.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"
flake8 = "^6.0.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
